Constrain missing value codes when replacing
The part of this function that replaces missing value codes with NA
was removing all entries when a period was the code. Now when periods
are used as the code, two slashes added to escape. Now, for all codes,
boundaries are added to the regex so a period doesn't cause "0.1" and
similar to be removed.